refactor: 음식점 리스트 정렬 및 필터링 명령 함수 위치 변경
